move warnIfExplicitCiBuildId
